# specs/001-haruhanjang-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 하루한장 API
  version: 0.1.0
  description: |
    사장님 PWA(S1~S8)와 기관 대시보드(S7)가 쓰는 REST API.
    근거: spec.md FR-xxx, design/SD_01 IPO, SD_02 화면, SD_03 테이블·뷰.

    공통 규칙
    - 모든 경로는 `/api` 아래. 인증은 HttpOnly 쿠키(`hrh_session`, 서명 JWT).
    - 날짜는 KST 기준 `YYYY-MM-DD`.
    - 게이트 차단은 HTTP 409 + `GateBlocked` 본문. 화면은 `gate` 코드로 SD_02 §11-2 표준 문구를 고른다.
    - 사장님 API는 로그인한 계정의 가게 데이터만 다룬다(FR-091). 경로에 store_id를 받지 않는다.
    - 기관 API(`/api/org/*`)는 익명 집계 뷰만 조회한다(FR-076, research R7).
    - 인가(RBAC): 각 경로의 권한은 data-model §7-2. 권한이 없으면 403 Forbidden, 게이트 차단은 409.
servers:
  - url: /api
...
paths:
...
  /alerts:
    get:
      tags: [alerts]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [new, acknowledged] } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Alert' } }
    post:
      tags: [alerts]
      summary: 경보 생성 (P2 2.9) — 기기 내 판정 결과 제출, 서버는 G3 재검증 후 저장·푸시
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlertInput' }
      responses:
        '201':
          description: 생성
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Alert' }
        '409': { $ref: '#/components/responses/GateBlocked' }
  /alerts/{alertId}/ack:
    post:
      tags: [alerts]
      summary: 경보 확인 (P3 3.2)
      parameters:
        - { in: path, name: alertId, required: true, schema: { type: integer } }
      responses:
        '204': { description: '확인함' }
...
  /paid/payments:
    post:
      tags: [paid]
      summary: 결제 시도 기록 (P6 6.3) — 공급자 미정, 개발은 모의 공급자
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [featureCode, providerResult]
              properties:
                featureCode: { type: string, enum: [report, backup_export] }
                providerResult: { type: object, description: 공급자별 결과 페이로드 }
      responses:
        '201': { description: '성공 — 권한 생성' }
        '409': { $ref: '#/components/responses/GateBlocked' }
  /paid/exports:
    post:
      tags: [paid]
      summary: 보고서·내보내기 생성 (P6 6.4) — 클라이언트 계산 스냅숏 제출, 서버는 G6·G3 재검증
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportInput' }
      responses:
        '201':
          description: generated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Export' }
        '409': { $ref: '#/components/responses/GateBlocked' }
  /paid/exports/{exportId}/confirm:
    post:
      tags: [paid]
      summary: 포함 정보 확인 (G7 해제)
      parameters:
        - { in: path, name: exportId, required: true, schema: { type: integer } }
      responses:
        '204': { description: 'confirmed_at 기록' }
  /paid/exports/{exportId}/file:
    get:
      tags: [paid]
      summary: 파일 받기 — confirmed 가 아니면 409 G7, 성공 시 delivered 로 전환
      parameters:
        - { in: path, name: exportId, required: true, schema: { type: integer } }
      responses:
        '200': { description: '파일 본문' }
        '409': { $ref: '#/components/responses/GateBlocked' }
  /paid/exports/{exportId}/decline:
    post:
      tags: [paid]
      parameters:
        - { in: path, name: exportId, required: true, schema: { type: integer } }
      responses:
        '204': { description: 'declined' }
...
  /push/subscriptions:
    post:
      tags: [push]
      summary: Web Push 구독 등록 (research R6)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, keys]
              properties:
                endpoint: { type: string }
                keys: { type: object, required: [p256dh, auth], properties: { p256dh: { type: string }, auth: { type: string } } }
      responses:
        '201': { description: '등록' }
    delete:
      tags: [push]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [endpoint], properties: { endpoint: { type: string } } }
      responses:
        '204': { description: '해제' }
...
components:
  responses:
...
    GateBlocked:
      description: 게이트 차단 (SD_02 §11, SD_03 §13)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GateBlocked' }
  schemas:
...
    GateBlocked:
      type: object
      required: [gate, reasonKey, releasedBy]
      properties:
        gate: { type: string, enum: [G0, G1, G2, G3, G4, G5, G6, G7, G8] }
        reasonKey: { type: string, description: 'SD_02 §11-2 표준 문구 키 (예: g3.reason)' }
        releasedBy: { type: string, enum: [self, others], description: 'others 면 화면에 해제 버튼을 두지 않는다 (FR-002)' }
        detail:
          type: object
          description: '문구 치환 값 (예: G3 recordCount, minRecords). 가게 수 등 익명성 수치는 넣지 않는다'
          additionalProperties: true
...
    AlertInput:
      type: object
      required: [windowStart, windowEnd, trendCode, confidenceLevel]
      properties:
        windowStart: { type: string, format: date }
        windowEnd: { type: string, format: date }
        trendCode: { type: string, enum: [decline] }
        confidenceLevel: { type: string, enum: [high, medium, low] }
    Alert:
      type: object
      properties:
        alertId: { type: integer }
        windowStart: { type: string, format: date }
        windowEnd: { type: string, format: date }
        trendCode: { type: string }
        confidenceLevel: { type: string, enum: [high, medium, low] }
        displayChannel: { type: string, enum: [push, in_app] }
        status: { type: string, enum: [new, acknowledged] }
        createdAt: { type: string, format: date-time }
...
    ExportInput:
      type: object
      required: [exportKind, recordCountSnapshot]
      properties:
        exportKind: { type: string, enum: [report, backup, export] }
        periodStart: { type: string, format: date, nullable: true }
        periodEnd: { type: string, format: date, nullable: true }
        recordCountSnapshot: { type: integer, minimum: 0 }
        confidenceSnapshot: { type: string, enum: [high, medium, low], nullable: true }
        reportBody: { type: object, description: 클라이언트 분석 모듈이 만든 보고서 내용(경향 문장·수치) }
    Export:
      type: object
      properties:
        exportId: { type: integer }
        exportKind: { type: string }
        periodStart: { type: string, format: date, nullable: true }
        periodEnd: { type: string, format: date, nullable: true }
        recordCountSnapshot: { type: integer }
        confidenceSnapshot: { type: string, nullable: true }
        status: { type: string, enum: [generated, delivered, declined] }
        includedItems: { type: array, items: { type: string }, description: 종류별 고정 목록 (G7 확인 화면) }